Move status_panel.py into RimSort.views
