+++ RimSort/views/status_panel.py
